[audio] Make dtor virtual
ASan is cacthing a new-delete-type-mismatch. We've allocated a
UltrasoundCapturerShim that gets deleted via a unique_ptr for a
CapturerShimImpl, but the child UltrasoundCapturerShim dtor isn't
properly invoked, only the parent CapturerShimImpl dtor is invoked. The
fix is to make the CapturerShimImpl dtor virtual so it is destroyed
properly.

Bug: b/354952331
Change-Id: I15e8377912eb0a54c20ad8df97e0beeaaff56b0b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1088534
Reviewed-by: Petr Hosek <phosek@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>
Owners-Override: Adam Barth <abarth@google.com>