[superproject] Roll llvm-project/libc [libc] Only add '-fno-builtin-*' on the entrypoints that use them (#100481)
Summary:
The GPU build needs to be able to inline stuff in LTO. Builtin
transformations cause problems on the functions that the optimizer does
heavy libcall recognition on. Previously we moved to using
`-fno-builtin-*` to allow us to only disable the problematic ones.
However, this still didn't allow inlining because each function had the
attribute that told the inliner not to inlining a nobuiltin function
into a non-nobuiltin function

This patch fixes that by only applying these attributes to the
entrypoints that define them. That is enough to prevent recursive calls
within the definitoins themselves.

GitOrigin-RevId: 604a6d9a5610ff056d430b07a0dfe69fb82e12a6
Original-Revision: 9a8ec5aaa02ba9ffc23525521a1f03a5005fd27b
Roller-URL: https://ci.chromium.org/b/8741465049350691985
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ibace2447837bdf1be0d01580e83a36ac4c7e3ad4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1088740