[cm_fidl_analyzer] include capability source in route results
The detailed route segments is going to be removed by future commits as
part of component framework's bedrock workstream. The most important
thing that these route segments are used to confirm is the source of a
route. This commit adds the capability source to route results, so that
tests can assert that the source of a route matches expectations without
using the route segments.

Change-Id: I917c6d33126bc85f2617c5993b25b04594372909
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1084412
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Laura Peskin <pesk@google.com>
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>