[ui] ActivityManager uses LeaseHolder
This change introduces a new struct `LeaseHolder` to handle wake lease
creation, holding, and destruction. For now this is a skeleton class
that does not take a real wake lease from the power framework. It will
be integrated with the power framework in a separate change after
fxrev.dev/1083132 is merged.

This change also parameterizes existing unit tests on the boolean
`suspend_enabled` and adds expectations around whether a "wake lease" is
being held in sync with activity state changes.

Bug: 343256679
Test: fx test activity-manager-test
Test: fx test input_pipeline_lib_tests --test-filter=*activity* -o
Change-Id: If18ff3866db3951907a75a1041f523f6bb43e4ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087514
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Commit-Queue: Caroline Liu <carolineliu@google.com>