[fx][build] Pass explicit job count to Bazel (again).
This is a reland of the reverted CL at [1] which fixes the bug
in `bazel_test.py` which caused the revert.

This CL ensures that if an explicit job count is passed to `fx build`,
then the corresponding value is passed both to Ninja _and_ Bazel at
the same time (previously, it was only passed to Ninja).

If no explicit job count is specified, the default behavior is
unchanged.

This can be useful to work-around Bazel spawning too many processes
at once on machines with limited amount of RAM compared to their
number of cores (see associated bug for an example).

- tools/devshell/build: Ensure using `fx build -j<count>` or
  `fx build -j <count>` does not print a warning anymore telling
  the user to pass Ninja-specific arguments after `--`.

- tools/devshell/lib/vars.sh: Properly parse the load and job
  counts arguments in fx-run-ninja, and set the
  FUCHSIA_BAZEL_JOB_COUNT environment variable when an
  explicit job count is provided. There is nothing that can be
  done with the load average value for now though.

- bazel_action.py, bazel_test.py: Update the launcher scripts
  to add `--jobs=<count>` when invoking Bazel, if
  FUCHSIA_BAZEL_JOB_COUNT is set in the environment.

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/1087015

Bug: 351623259
Bug: 353989247
Change-Id: Ic09835327e146ebf2a699630f72e1816462982a9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089135
Reviewed-by: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>