[kernel][phys] Remove kArchHandoffGenerateUartPeripheralRanges logic
This is used to synthesize a PERIPHERAL range to cover UART MMIO;
however, this is now redundant with similar logic added to the arm64
Linux devicetree boot shim and first-party bootloaders being required to
supply it to cover all kernel MMIO access. This also is applied to the
construction of `PhysHandoff::mem_config`, which is slated to be
replaced in the context of fxbug.dev/347766366, so it seems prudent to
remove it instead of look to add new support for it in the new codepath.

Bug: 347766366
Change-Id: I8bd058bb2070292620913134cbf250deab9d52f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089237
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>