[blobfs] Immediately send decompression request
Don't wait for the fifo to be writable before sending a decompression
request. Instead, try writing to the fifo and then handle any write
errors. There's a decompression thread for each pager thread and the
pager threads wait for the decompression response before sending another
request. It shouldn't be possible for requests to back up in the fifo,
the fifo should always be writable.

Bug: b/350020539
Change-Id: Id553e2cf21df74bdb542c6f749bd321685ad2899
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1079217
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>