[starnix] Only run gvisor tests on hardware.
There are a number of tests we inherit from gvisor which use
sleep() to "control" the ordering of the test logic between multiple
processes. It seems that these tests sometimes flake on emulators
because the underlying host has not scheduled the emulator recently
enough.

Move these tests to only running on hardware to get more predictable
timing.

In the future, we may be able to work with the gVisor team to migrate
to other methods of synchronization but for now this should reduce
the flake rate.

Needs internal CL 864671.

Bug: 355311252
Bug: 353614027
Fixed: 354249283
Change-Id: I9f9ba653285c6db869f6fe74d060330dd1b11da2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1084721
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Reviewed-by: Vickie Cheng <vickiecheng@google.com>