[network/test] De-flake DNS timeout test
The flaking test case exercises the DNS resolver's 5s query timeout, and
the behavior that it will return records of one type (e.g. AAAA) even if
it doesn't receive any of another type it requested (e.g. A). The way
it's written is timing-dependent because it advances a fake clock for
the DNS resolver by 1 second every 10 milliseconds of test runtime. This
means in rare cases, even the record that *is* returned by the fake name
server doesn't arrive and get processed in time before the 5s timeout.
Increase the ratio of test runtime to fake DNS resolver time so the
response is more likely to be processed.

Fixed: 355083806
Multiply: netstack-dns-integration-test
Change-Id: I27c0c6e51d3b0131e7bbc0d2e2142c0475e69000
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089512
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>