[fxfs] Tombstone to not fetch Object mutation unless needed.
This change is mean to support a change to the find() interface where we
do not return deletion placeholder values. It does not fix the
underlying issues regarding `intial_reap()` but does make it more robust
in being idempotent when deleting attributes that do not exist.

In short I'm papering over the issue differently, to support the
interface change until we can properly fix the caller.

Bug: b/355246066 b/355258430

Change-Id: I65173dbbeff7972069d85aef2e7333f74bdbe6df
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087735
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Chris Drouillard <cdrllrd@google.com>