[kernel][phys] Hack PhysLog to leak memory
While trampoline booting is in effect in the x86 codepath, care needs to
be taken with `Allocation`s made before the fixed-address image
recharacterization that TrampolineBoot does. In particular, we do not
want to absent-mindedly free a region (e.g., due to Allocation::Resize()
or Allocation::~Allocation()`) that was recharacterized as being  a part
of a fixed-address image. Until TrampolineBoot is removed from kernel
boot, defensively leak log buffer allocations when trying to resize or
destruct an Allocation.

Bug: 347766366
Change-Id: I8cfc705baaea43bab4069cccb95489b76aa956ed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1088739
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>