[shac] Settle the age-old question: _ or -?
The answer is: do whatever files near you (with the same extension as
you) do. This is a SHAC check that attempts to codify that answer.

It only ever emits a warning, which we could also downgrade to a notice.
I manually looked through exists files that are out-of-compliance, and
they largely seem to fit the category of "files that would ideally be
differently named."

Change-Id: I65b8afd47684d3cd7e3f4327a1d6fb8ed22429c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1082053
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Hunter Freyer <hjfreyer@google.com>