[starnix] Add diagnostic checks for TheadGroup children
This forces `remove` to takes a OwnedRef<Task> to ensure that the task
is still alive when it is removed, and this add a debug only check on
ThreadGroup that the children map is empty when the ThreadGroup is
removed.

The goal is to get better information on why there is case where some
children cannot be upgraded.

Bug: b/354754795
Change-Id: I6ca617aaa55f867bbb8a34e2fce30177f2f67b86
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089834
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>