[netstack3] Rework multicast route witness types
Rework the witness types of source addresses for multicast routes. The
source addresses must be unicast, but `UnicastAddr<Ipv4Addr>` is not a
safely constructable type (because unicastness of an IPv4 address can
only be determined in the context of a subnet).

For IPv4, rather than using `UnicastAddr`, use a custom witness type
that ensures "specified and not multicast".

Bug: 353329136
Change-Id: I93f655a3ea9d1704e6616beb9a52de48e9d31dd8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089036
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>