[fxfs] LSMTree.find() to stop returning an item when value is None.
This requires adding a known placeholder type for the Value trait so
that the None value can be identified and filtered in find(). This makes
sense for all production uses, though some of the test use cases where
we lazily use ints are strange, but they're not a critical case anyways.

Bug: b/355258430

Change-Id: Icbdea56607343adb4bcc453a70945d08294da934
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087737
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>