[fake-display-stack] Add SysmemServiceProvider interface.
Previously, `FakeDisplayStack` created a mock-ddk-based device root
and used the `GenericSysmemWrapper` template class to bind the sysmem
"devices" to the root. This caused a lot of confusion in the fake
display stack:

(1) the SysmemProxyDevice doesn't need to be bound to the fake root;
(2) the template class requires the SysmemProxyDevice and
    sysmem_driver::Device to have similar interfaces, which is also
    not necessary;
(3) SysmemProxyDevice implements the DriverConnector protocol, which
    is not usable by any clients.

This change is the first step to refactor the FakeDisplayStack.

This change defines a new interface type `SysmemServiceProvider`. Its
only responsibility is to provides the fuchsia.hardware.sysmem/Service
to an outgoing directory, consumed by display Coordinator and the
fake-display library.

The `SysmemServiceProvider` doesn't place any constraints on the source
of the sysmem service protocols. It could be from the component's
incoming directory, or from a self-owned fake sysmem device.

Test: no-op
Bug: 354836088
Change-Id: Ic77e291ccb972b78064c2fdf13731ae21f6febcd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087592
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>