[fidlc] Add better error for name not found with versioning
This improves the fidlc error message when a name is not found in a
versioned library. Before, it would just say "name not found". Now, if
the name actually is found but in the wrong version range, it lists all
the version ranges and locations it did find the name.

Test: fx test fidlc-test
Change-Id: I9d02789dc84279a0e682e41ac5635d723c04a6bd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1077906
Fuchsia-Auto-Submit: Mitchell Kember <mkember@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>