[fake-display-stack] Add SysmemServiceForwarder.
This change adds SysmemServiceForwarder which is an
implementation of the `SysmemServiceProvider` interface.

The SysmemServiceForwarder is similar to the
SysmemProxyDevice but it is not a device type and it only
forwards the sysmem protocols to its outgoing service directory.

Test: component-sysmem-service-forwarder-tests
Bug: 354836088
Change-Id: Idaca8e338b1458f3bde055dbd85ba56b24e5e817
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087593
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>