[netstack3/conntrack/tcp] SynSent
Since transitioning to the next state requires knowing all of the data
it will contain, this also adds the Peer struct, which will be core to
all the upcoming states (minus TimeWait).

Implementing this state took a bit of refactoring to plumb the new
update status result into the right places, but no more should be
required for the remaining states.

Bug: 328064736
Change-Id: I204c4dc11557c9501b843c4d03326416ac7dc77f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1088372
Commit-Queue: Peter DeLong <pdelong@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>