[ffx][build] Adding --no-environment to ffx_action()
This adds the `--no-environment` switch to the
ffx_action() template by default.

This is done to improve hermeticity of running ffx during
the build.

The option is exposed as a parameter in the template if there
is some edge-case that requires the configuration to be read.

Bug: b/355672251
Change-Id: Ie9ac10a6c07d8fe37a2440ec8cfe9a710f71fbc3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1090795
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>