[fx][rust3p] Call ninja directly instead of `fx build`
This removes the warning from `fx build` that directs users to use GN
targets instead of Ninja outputs.

I opt to not use GN target for build in this particular case because
it doesn't improve the script. In fact it makes it slightly worse
because moving the GN target would requiring changing this script, which
building the Ninja output directly does not, and we'll need to hardcode
the path to the Ninja output anyways to invoke it.

Change-Id: Iaf93fc5257990730a3470e7dc7df78d53f61fa11
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089714
Fuchsia-Auto-Submit: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Turner <digit@google.com>