[msd-arm-mali] Add magma info dump to inspect
Having information about the state of the driver is useful in hang
scenarios. A lot of the scheduler code is on a latency-sensitive path,
and some information is read on demand through hardware registers, so we
can't update it as changes happens.

Because of that, it makes the most sense to create a LazyNode that grabs
the GPU dump whenever it's requested.

Bug: 355485837
Change-Id: If622809483d1575f91ce6f1dad1d37896a4b517e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1090032
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>
Reviewed-by: Craig Stout <cstout@google.com>