[ffx][pkg] Support for managing server process instances
This adds a struct to manage instance files that contain
the information used to start server processes.

This information is used to implement stop and list commands
and could be used to identify the appropriate package server to
use for a command based on the project and the target device.

Bug: b/342205668
Change-Id: I2b210cfc476e90b4cc1b516fb992a62eaddc8cb1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1051464
Reviewed-by: Thomas Zander <thomasalpinus@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>