[sys/test_manager] integrate debug_agent
This CL integrates debug_agent into the test environment to enable
the generation of backtrace files in the test output when tests fail
or time out. For the time being, this feature is disabled due to issues
with exception channel conflicts. Once those conflicts are resolved,
another CL will enable this feature.

When enabled, backtraces go to a stderr file for the suite, which is
shown symbolized in the report that goes to the console when running
'fx test'. Using the stderr file instead of a special file allows the
backtrace to be displayed without host-side changes. A future CL will
introduce a new artifact type for this purpose.

This CL causes Rust backtraces to appear twice in the fx test console
output, because Rust backtraces are delivered using a different
mechanism. This issue will be addressed separately.

Bug: 286555768 303314982

Test: unit tests added

Change-Id: If34e31304cace8b336207eb82dba3f945d33667f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1038975
Reviewed-by: Ankur Mittal <anmittal@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>