[ffx] Fixes a bug with Fastboot mDNS discovery.
The fastboot mDNS discovery code would only check a single record to
determine if the domain is correct, rather than all records. This would
lead to flakes where the record we were interested in would not be
detected.

This also adds mdns_discovery to the list of tested libraries in ffx.
For some reason it was not included so likely wasn't being tested in CQ.

Bug: b/353345798
Change-Id: Ib15c4bb1d8eaa8f562c62c3d29d4ea8c857bee6f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089926
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>