[pkvm] Ship pkvm boards in the minimal builders
The pkvm-x64 and pkvm-arm64 boards are now built and uploaded to CIPD in
continuous integration.

In order to ensure unique board names, we have to override the `name`.
The board_configuration.gni had to be modified to use the overridden
name instead of the target_name.

Change-Id: I72397dfc6e59bda88ae12d2818624868418099ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1090806
Reviewed-by: James Robinson <jamesr@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>