[broker] Preserve dependencies during transition
On transitions between power levels, elements may be in either state and
the broker should attempt to defer decisions to activate or deactivate
claims until the state is well-known. The broker now checks to see if a
claim is satisfied only if it is satisfied at both it's current level
and it's transition level. It also does not drop claims that are
currently transiting.

Further, this change does not send required level updates until the
element replies to a previous required level change. This ensures that
only one transition is ongoing for an element at any given time.

Bug: 342450114
Test: fx test src/power/broker
Change-Id: I3b933fa558d5565f131d4173f47d15a7f9a34e22
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065215
Fuchsia-Auto-Submit: Prashanth Swaminathan <prashanthsw@google.com>
Reviewed-by: Thomas Gales <tgales@google.com>
Reviewed-by: Michael Brunson <mbrunson@google.com>
Reviewed-by: Guocheng Wei <guochengwei@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Onath Dillinger <claridge@google.com>