[netstack3] De-macro "process_fragment"
Convert the "process_fragment" macro into a function. Originally it was
a macro to circumvent problems with the borrow checker: It was taking in
both `buffer` and `packet` as arguments, which as a function call would
"move" `buffer` and invalidate the reference held by `packet`.

Here, I move error handling to outside of the `process_fragment` call,
which means it no longer needs to accept `buffer` as an argument, and we
can avoid reference invalidation.

Change-Id: I8a001052a6d2943248e5904a015de36d284cc364
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1091573
Reviewed-by: Peter DeLong <pdelong@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Jeff Martin <martinjeffrey@google.com>