[starnix] Trace the start and completion of suspend
This was requested in comment #12 of b/348021898. The motivation is to
understand the performance of these operations.

Test: I ran

  ffx trace start --categories "starnix_runner" --buffer-size 64 --duration 10

on one terminal, and

  ffx starnix suspend
  ffx starnix resume

on another terminal. I was able to find suspend durations in the
resulting `fxt` file. The duration slices may be very hard to spot if
your container is not doing much.

Bug: b/348021898
Change-Id: Id16c4fd93da190072b61b1c997fe24836f5c2d02
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1091338
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Adam Barth <abarth@google.com>