[kernel][phys] Ensure the correct allocations will survive into hand-off
There are two cases around BootZbi::Load() in which we were ending up
with duplicate allocations with type kKernel and kDataZbi:

(1) In physboot, we Load() a kernel ZBI alone, yielding a spurious data
ZBI allocation from its empty tail, which we then overwrite.
(2) TrampolineBoot allocates staging kernel and data ZBI images for the
purposes of trampolining, but these clash with the fixed-address
versions we reserve ahead of time.

We expediently address both of these situations with simple freeing or
recharacterization of types post-Load(). This is easy enough to do,
readable, and in (1) involves codepaths that are not long for this
world.

As logging feedback that the correct kKernel and kDataZbi allocations do
end up surviving, we add these types to those included in the desired
set of hand-off allocations.

Bug: 347766366
Change-Id: I2f801967ae936ea9a3584736443a5012dcb89e42
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1081985
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>