[wlan-frame-writer] Split Appendable into Append and TrackedAppend
The motivation for this change is that the implementation of
Appendable::bytes_written for a Vec<u8> was ambiguous. As implemented,
the return value was always the length of the Vec<u8> regardless of
how many bytes had actually been appended to it.

To resolve this ambiguity, this changes does the following:

  - Rename Appendable to Append.
  - Move Appendable::bytes_written to TrackedAppend::bytes_appended.
  - Creates VecCursor.
  - Renames write_frame_with_appendable_buffer!() to append_frame_to!().
  - Creates write_frame_to_vec!().

The effect of these changes is that implementations of Append now
do not need to support a bytes_appended method. Instead,
bytes_appended is only provided by TrackedAppend types.

To improve ergonomics, this changes creates write_frame_to_vec!()
which prevents needing to create an empty Vec in many use cases.

This change is a bit long-winded, but my hope is that it will prevent
misuse of types that implement Append in the future.

Change-Id: I7dc443966175f7b970358922ae3605e34ebf49a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1088734
Commit-Queue: Charles Celerier <chcl@google.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Reviewed-by: Sean Olson <seanolson@google.com>