[fake-display-stack] Add FakeSysmemDeviceHierarchy.
This change adds FakeSysmemDeviceHierarchy which is an
implementation of the `SysmemServiceProvider` interface.
It creates a mock device hierarchy with a fake sysmem
Device.

This is similar to the GenericSysmemWrapper with a
fake "sysmem_driver::Device". The difference is that the
FakeSysmemDeviceHierarchy contains all the dependency
devices and doesn't depend on any components from the
FakeDisplayStack.

Test: fake-sysmem-device-hierarchy-tests
Bug: 354836088
Change-Id: I8d5883936773c95ee4b29cfa13ec6ee004d4e240
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087594
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>