[ktrace] Force IPC Flows to be Correctly Ordered.
With the current IPC events, if we are too fast between creating the
duration scope and emitting the flow, we may get a series of trace
events which look like:

```
{
  "cat": "kernel:ipc",
  "name": "ChannelFlow",
  "ts": 3706186824.6666665,
  "pid": 4985, "tid": 4987,
  "id": "15440798007190264229"
}
{
  "cat": "kernel:ipc",
  "name": "ChannelMessage",
  "ts": 3706186824.458333,
  "pid": 4985, "tid": 4987,
  "args": {
    "method": "fuchsia.power.system/ActivityGovernorListener.OnResume"
  }
}
{
  "cat": "kernel:ipc",
  "name": "ChannelFlow",
  "ts": 3706186850.958333,
  "pid": 4985, "tid": 4987,
  "id": "948362557191958181"
}
{
  "cat": "kernel:ipc",
  "name": "ChannelMessage",
  "ts": 3706186850.958333, // OOPS: Same timestamp as the previous event
  "pid": 4985, "tid": 4987,
  "args": {
    "method": "fuchsia.power.broker/CurrentLevel.Update"
  }
}
```

Because we use complete events, the duration is emitted after the flow.
This is usually fine because we can consult the timestamps to figure out
when the duration started. However, because creating a duration complete
scope doesn't actually write any data, its possible to create the
complete scope and start string the flow event without the timestamp
increasing. This causes it to be ambiguous which event started first.
Perfetto defaults to whichever event comes first in the trace is the
event which comes first, which causes the flow event to be associated
with the first flow rather than the second.

This CL forces the ordering in the trace of the events so that the flow
will always be contained in the duration. This also has the positive
side effect that these zero-length durations show up in perfetto as
instant events which are much easier to click on.

This does however, have the negative of emitting slightly more trace
data: an increase from 96 bytes to 120 bytes (24 bytes) for each flow.

In addition, added *_TIMESTAMP trace event variants that accept an
expression to evaluate for the timestamp of the event. This allows the
caller to customize the capture and reporting of timestamps for better
control over visualization of events with specific relationships.

Then, use the new _TIMESTAMP trace event variants to fix the timestamps
of channel message flow events. Also avoid sampling the timestamp
counter unconditionally, which could incur non-trivial overhead on some
architectures and virtualized environments.

Testing:

Took a long trace with kernel:meta,kernel:sched,kernel:ipc enabled:
'https://ui.perfetto.dev/#!/?s=d1323c6b20917e9b3ed2e9351f0ffbb552a7d2ef'

Querying for flows that have a different from and to fidl method returns
0 results.

```
DROP TABLE IF EXISTS my_thread_info;
CREATE TEMPORARY TABLE my_thread_info as
  SELECT
    thread_track.id as track_id,
    thread_track.utid,
    thread.tid,
    thread.name as thread_name,
    process.pid,
    process.name as process_name
  FROM thread_track
  INNER JOIN thread on thread_track.utid = thread.utid
  INNER JOIN process on thread.upid = process.upid;

SELECT
  COUNT(flow.id) as fidl_count,
  SUM(slice_in.ts - slice_out.ts) as total_duration,
  AVG(slice_in.ts - slice_out.ts) as avg_duration,
  args_out.string_value as fild_method,  args_in.string_value as fild_method_2,

  thread_info_out.tid as from_tid,
  thread_info_out.thread_name as from_thread_name,
  thread_info_out.pid as from_pid,
  thread_info_out.process_name as from_process_name,
  thread_info_in.tid as to_tid,
  thread_info_in.thread_name as to_thread_name,
  thread_info_in.pid as to_pid,
  thread_info_in.process_name as to_process_name
FROM flow
  INNER JOIN slice as slice_out on slice_out.id = flow.slice_out
  INNER JOIN args as args_out on slice_out.arg_set_id = args_out.arg_set_id
  INNER JOIN my_thread_info as thread_info_out on slice_out.track_id = thread_info_out.track_id
  INNER JOIN slice as slice_in on slice_in.id = flow.slice_in
  INNER JOIN args as args_in on slice_in.arg_set_id = args_in.arg_set_id
  INNER JOIN my_thread_info as thread_info_in on slice_in.track_id = thread_info_in.track_id
WHERE
  slice_in.name = "ChannelMessage"
  AND slice_out.name = "ChannelMessage"
  AND args_in.string_value <> args_out.string_value
GROUP BY
  args_out.string_value,
  args_in.string_value,
  slice_in.track_id,
  slice_out.track_id;
```

Co-authored-by: eieio@google.com
Fixed: 355282759
Change-Id: I4c9deb3a5ec998b04727bd4f8c78143eb1dc48ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1088855
Reviewed-by: Fadi Meawad <fmeawad@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>