[bt][gap] Write Variable PIN Type for Legacy Pairing
A BR/EDR connection using legacy pairing must use a PIN code that is
either a variable or fixed type. Sapphire chooses to support variable
PINs, so we write this information the controller when
BrEdrConnectionManager is created.

This change is part of the larger effort to support BR/EDR Legacy
Pairing.

Bug: b/42173830, b/342151162
Test: fx test bt-host-gap-tests
Change-Id: Id389c725977ee30c2b263bd7cc267630a4caf9f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1086961
Commit-Queue: Lulu Wang <luluwang@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>