[kernel][x86] Use memmove for copying kaslr-fixups.
The code that emits the relocation fixups for kaslr is copied out from
the bss section during `_start`, into the location that's right after
the bss section.

If the kernel contains too many relocations, as is the case for some
kernel variants like `kubsan`, the relocations will partially overlap
the destination buffer. This caused an issue with the previous code for
copying out the relocations: by copying from left to right, it would
overwrite the same code it is copying.

The solution is to copy right-to-left if the destination buffer overlaps
with the source buffer.

Bug: 328825932
Change-Id: Ief665e4337c692ce7dad885d8520586eb6521361
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1068314
Reviewed-by: Leonard Chan <leonardchan@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>