[test] Updating device enumeration test
This removes hardcoded node monikers that is not
created when running a different configuration of AEMU.

This was discovered when changing infra test execution from a custom
configuration to use the standard configuration used by developers.

There is follow-up work planned to improve device enumeration to
make the names easier to verify.

Bug: b/329144967
Change-Id: I7b3fc5936b502177e1086a0a8b5db4e2545e7a73
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1092100
Reviewed-by: Sarah Chan <spqchan@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>