[routing] remove capability_provider from FilteredAggregate
`CapabilitySource::FilteredAggregate` was the only remaining
(non-phantom) usage of generics in `CapabilitySource`. This was due to
the `capability_provider` field, a `Box<dyn>` of a trait defined in the
routing library.

There were only two different values ever placed in this field:
`OfferFilteredServiceProvider` and `OfferAggregateServiceProvider`. This
commit removes `CapabilitySource::FilteredAggregate` and replaces it
with two different variants. `CapabilitySource::FilteredProvider` holds
the information necessary to construct an
`OfferFilteredServiceProvider`, and
`CapabilitySource::FilteredAggregateProvider` holds the information
necessary to construct an `OfferAggregateServiceProvider`. With this
change, there are no remaining (non-phantom) instances of generics in
`CaapbilitySource`.

Change-Id: I92f4d9f433f8c7e83c742cdbe1c73b245abd2d89
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1090595
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>
Reviewed-by: Gary Bressler <geb@google.com>