[starnix] Propagate EAGAIN from std::thread::Builder::build
Currently starnix panics if std::thread::Builder::build fails, but it
can fail with EAGAIN when thread creation fails due to insufficient
resources. Propagate this particular error variant to userspace.

This requires careful handling to ensure that the TaskBuilder provided
to the spawned thread is always explicitly released; we introduce a
one-shot channel to provide the builder to the thread only in the case
that the thread was successfully spawned.

Change-Id: Id7484257f2d881fd46126fa294f074011959de9e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1086597
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: Benjamin Lerman <qsr@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>